feat(navigation): update navigation links to reflect GameDesign as page title
Adjusted the navigation links in `_data/navigation.yml` to replace references to "DevLog" with "GameDesign" in various language configurations. Also, changed the URL from `/DevLog.html` to `/GameDesign.html` for consistency with the new page title. Added a new file `GameDesign.html` with metadata defining it as a home layout containing articles as HTML excerpts.